requirements.txt
# Langchain & LangGraph
langgraph
langchain
langchain-openai
langchain-pinecone
typing-extensions

# Kivy ve Bağımlılıkları
docutils
pygments
pypiwin32
kivy.deps.sdl2
kivy.deps.glew
kivy.deps.gstreamer
kivy.deps.angle
kivy

langdetect
python-dotenv